# apps/api/sdks/typescript/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Janua API",
    "description": "Modern authentication and identity platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/passkeys/register/options": {
      "post": {
        "tags": [
          "passkeys"
        ],
        "summary": "Get Registration Options",
        "description": "Get WebAuthn registration options",
        "operationId": "get_registration_options_api_v1_passkeys_register_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegisterOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/passkeys/register/verify": {
      "post": {
        "tags": [
          "passkeys"
        ],
        "summary": "Verify Registration",
        "description": "Verify WebAuthn registration",
        "operationId": "verify_registration_api_v1_passkeys_register_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/passkeys/authenticate/options": {
      "post": {
        "tags": [
          "passkeys"
        ],
        "summary": "Get Authentication Options",
        "description": "Get WebAuthn authentication options",
        "operationId": "get_authentication_options_api_v1_passkeys_authenticate_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyAuthOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/passkeys/authenticate/verify": {
      "post": {
        "tags": [
          "passkeys"
        ],
        "summary": "Verify Authentication",
        "description": "Verify WebAuthn authentication",
        "operationId": "verify_authentication_api_v1_passkeys_authenticate_verify_post",
        "parameters": [
          {
            "name": "challenge",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Challenge"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/passkeys/": {
      "get": {
        "tags": [
          "passkeys"
        ],
        "summary": "List Passkeys",
        "description": "List user's passkeys",
        "operationId": "list_passkeys_api_v1_passkeys__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PasskeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Passkeys Api V1 Passkeys  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/passkeys/{passkey_id}": {
      "patch": {
        "tags": [
          "passkeys"
        ],
        "summary": "Update Passkey",
        "description": "Update passkey name",
        "operationId": "update_passkey_api_v1_passkeys__passkey_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "passkey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Passkey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "passkeys"
        ],
        "summary": "Delete Passkey",
        "description": "Delete a passkey",
        "operationId": "delete_passkey_api_v1_passkeys__passkey_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "passkey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Passkey Id"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasskeyAuthOptionsRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "title": "PasskeyAuthOptionsRequest",
        "description": "Passkey authentication options request"
      },
      "PasskeyAuthRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "credential": {
            "type": "object",
            "title": "Credential"
          }
        },
        "type": "object",
        "required": [
          "credential"
        ],
        "title": "PasskeyAuthRequest",
        "description": "Passkey authentication verification request"
      },
      "PasskeyRegisterOptionsRequest": {
        "properties": {
          "authenticator_attachment": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(platform|cross-platform)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authenticator Attachment"
          }
        },
        "type": "object",
        "title": "PasskeyRegisterOptionsRequest",
        "description": "Passkey registration options request"
      },
      "PasskeyRegisterRequest": {
        "properties": {
          "credential": {
            "type": "object",
            "title": "Credential"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "credential"
        ],
        "title": "PasskeyRegisterRequest",
        "description": "Passkey registration verification request"
      },
      "PasskeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "authenticator_attachment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authenticator Attachment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "sign_count": {
            "type": "integer",
            "title": "Sign Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "authenticator_attachment",
          "created_at",
          "last_used_at",
          "sign_count"
        ],
        "title": "PasskeyResponse",
        "description": "Passkey response model"
      },
      "PasskeyUpdateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PasskeyUpdateRequest",
        "description": "Passkey update request"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}